Remove all methods from dayOffRepository
